requirements.txt
django
django-rest-framework
loguru